highest normalized range for a specific day tests

+++ src/main/java/com/investment/dto/CryptoCurrencyRecord.java
@@ -2,5 +2,5 @@ package com.investment.dto;
 
 import java.time.LocalDateTime;
 
-public record CryptoCurrencyRecord(LocalDateTime date, float price) {
+public record CryptoCurrencyRecord(LocalDateTime date, double price) {
 }
